requirements.txt
absl-py==0.11.0
appdirs==1.4.3
astor==0.8.1
CacheControl==0.12.6
cached-property==1.5.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
html5lib==1.0.1
idna==2.8
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.4.0
ipaddr==2.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lockfile==0.12.2
Markdown==3.3.3
matplotlib==3.3.3
msgpack==0.6.2
networkx==2.5
numpy==1.19.5
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.3
pandas==1.1.5
pep517==0.8.2
Pillow==8.1.0
progress==1.5
protobuf==3.14.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.5
PyWavelets==1.1.1
requests==2.22.0
retrying==1.3.3
scikit-image==0.17.2
scipy==1.5.4
Shapely==1.7.1
six==1.14.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
tifffile==2020.9.3
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
